+++ NEWS.md
@@ -2,8 +2,8 @@
 
 ## Changes to functions
 
-* `icc()` gets `ci` and `iterations` arguments, to compute confidence intervals
-  for the ICC, based on bootstrapped sampling of ICC values.
+* `icc()` and `r2_nakagawa()` get `ci` and `iterations` arguments, to compute
+  confidence intervals for the ICC resp. R2, based on bootstrapped sampling.
 
 * The model underlying `check_distribution()` was now also trained to detect
   cauchy, half-cauchy and inverse-gamma distributions.